Rework the addPlayerMark function
In previous iteration, wrote a combined takeTurn function
for both players. As a result, created too many event
listeners on the square elements.

In this iteration, decide to divide the takeTurn into
two functions with the same code. Despite repetitive codes,
it's easier to read through the program and maintain in the
future. It doesn't create too many event listeners either.
Whoever turn it is, the program only adds the event listener
for that particular player.